Fix AddDelayed not running while the window is unfocused
We were using EditorApplication.delayCall for this, which doesn't run while the window is unfocused. It's fixed by using the UITK scheduler. This was very noticeable when working with two monitors: when the Editor was visible but unfocused, and a Domain Reload happened after changing a script, some elements would only appear after clicking on Unity.