Remove delay for ManagedReferenceFields reacting to editor changes
This is nice for UX, as the UI's response to user actions is more immediate, but that's not the main reason for this change.

Turns out UI Toolkit prints errors when a managed reference field changes class, and that class itself has a field with the same name as the previous one, but different type. To avoid it, we need to update the UI before Unity has a chance to trip up.

Previously, we were waiting for UIToolkit to run before we did, to avoid redundant SerializedObject updates, as UIToolkit itself updates the SerializedObject. With these changes, we've taken measures to avoid redundant SerializedObject updates as much as possible. These measures seem to work well enough.

Sadly, these changes don't prevent UIToolkit from printing errors when the change happens from runtime code. To solve that, we'd need to check for type changes with a lot more frequency, which would either be quite slow and generate a lot of garbage, or would require lots of reflection that would need to be different for different Unity versions. The scenarios where fixing this would be useful seem to be rare enough, so we are not doing it, at least for now.