Change implementation of EditableLabel
I tried to make it work in 2022, but those TextFields don't like it when they are disabled, enabled and focused like that. The code was becoming really ugly.

Editable Labels are now implemented using both a Label and a TextField. I wanted to avoid that because, before 2022, TextFields use a different text engine than labels; so it's harder to make them both fit exactly the same space. But making that work in 2022 is just not worth it, and I don't want to maintain different implementations for different versions of Unity.

Plus, before 2022, Labels handle zooming in and out much better; text fields disappear words when transforming their scale. This makes some UI look better before 2022, as EditableLabels are seen in their reading mode most of the time.

The API is mostly the same, but there are some breaking changes courtesy of hierarchy and styling changes. I was planning to release the next UITKEditorAid version as a 3.0 anyway.