Add a RemoveFromDocsAttribute to overridden members
It's the most practical way I found to remove those members from DocFx documentation.

In the vast majority of cases, users shouldn't care that a virtual member is overridden in a child class. Ideally, documentation for inherited members should show the same for overridden and non-overridden, as it has the same uses in both cases.