Add some fields to package.json
Adds documentationUrl, changelogUrl, and license.

License doesn't seem to do anything in the package manager, but we set it in case we missed something, or something is added in the future.

We use the GitHub releases page for the changelog. It contains all the the info that would normally go in a changelog, so it's good enough for this.