Remove a couple of TODOs in ListControlStyle
* Remove TODO about compensating for special field margins that Unity adds inside inspector elements. The note mentions PropertyFields, but it really affects all BaseFields.

Unity adds a negative right margin to those fields in the inspector; this makes them match IMGUI fields better in default conditions, but it messes up any other use. It's an objectively bad decision:
    -Things don't match between the Inspector and other windows or the UI Builder.
    -It makes layouts messy when fields are arranged side by side.
    -It doesn't work for any other element, e.g. buttons.
    -It messes things up when CustomEditors override UseDefaultMargins.
    -The same could've been accomplished without these problems by reducing the right padding of the default Inspector container.

I've decided the ListControl is not the place to fix this. Ideally, this should be fixed by Unity; I've posted about this in the forum, but I didn't have success. Hopefully, some notices and fixes this problem one day. Users can do the fix themselves with a stylesheet that sets the BaseField's right margin to 3px. I won't add this fix to ListControl because I don't want to risk overriding a margin set by the user in a stylesheet added higher up in the hierarchy. Maybe I can add an utility method for users to add this fix to any element they want.

* We also Remove a TODO about top and bottom margins in the header. It doesn't seem to be useful.