Fix EditableLabel editing not working on double-click  in 2023.2
Turns out 2023.3 requires calling focusController.IgnoreEvent to avoid removing the focus from the TextField. It seems now the focus handling happens at the end of an event's life-cycle, so Unity treats it like a click on the non-focusable EditableLabel after focusing the TextField, which causes the BlurEvent to be fired, stopping the editing.

Instead of calling IgnoreEvent, we decide to delay Focusing the field. This way, it works everywhere without needing to remember to call IgnoreEvent before calling BeginEditing.