Remove Disabler from ListOfInspectorHeaders
The idea was that the InspectorElements already handled disabling themselves when their Object wasn't editable due to Version Control or their hideFlags, so we just needed to also disable the header, as it contains a menu that could edit the Object, and a field to edit the name.

It turns out we were wrong. InspectorElement doesn't disable itself. In normal inspector windows, other internal element takes care of checking if the object is editable. Sometimes, in some versions of Unity, bound UITK fields are disabled when they aren't editable, even outside inspectors, but they don't seem to detect when they aren't editable because of Version Control.

I thought it worked, but it actually was because I only use ListOfInspectors with nested subassets in the inspector of a main asset, so Unity was disabling the whole main inspector when it needed an action from Version Control.

So disabling the header is kind of useless if we aren't disabling the inspector too. To do it well, we'd have to detect when the Object isn't editable in a separate place, and then disable only the parts that could edit the Object while allowing actions like toggling the foldout.

The thing is, that functionality is not very useful for my case. Whenever a single Object in the list should be disabled because of Version Control, the whole main Inspector is disabled because they are part of the same main asset. So I don't have any need to work on that. Maybe this functionality could be added later if it turns out someone needs it.