MAYBE BREAKING: Use attributes instead of UXML factories in v6
Unity is removing UxmlFactory and UxmlTraits. They are marked obsolete now, and they will be removed in the future. On the other hand, they are the only way of supporting UXML in older versions. There will come a time when we'll have to use one feature in older versions and another in newer ones, and we are deciding to go ahead and make that time now.

We could mark our factories obsolete and keep them in Unity 6 to avoid the breaking change, but the thing is these factories would only be obsolete for Unity 6 users. It's such a weird Limbo that would be followed by another weird Limbo where Factories don't exist only in newer Unity versions. I'd rather only have one weird Limbo in a new mayor release of this library; for one, it's less work.