Create ValuePropertyTracker, ValueTracker is now obsolete
ValuePropertyTracker handles quirks in the binding system related to mixed values. It uses a change callback that doesn't depend on UITK's event system, so it can be triggered when the value is updated before adding the element to a panel. It also has a callback for when the mixed value status changes.

We are creating a new class instead of updating ValueTracker with the new callbacks because it gives us a chance to make implement SetValueWithoutNotify explicitly. This is important because our callback is triggered even when the value is set using that method.