Make ListOfInspectors bind to the nested Editor's serializedObjects
It turns out that when InspectorElement is in charge of creating the Editor, it binds the UI to a different serializedObject than the Editor.serializedObject. For the main inspector, it's usually the same, so people might expect a custom Editor's serializedObject to be kept updated by UITK's binding system.

We fix this by creating and managing the InspectorElement's Editor on our own.