Changes to ListOfInspectors headers
We move the tooltip to the Label. It's easier to discover and it's less confusing when there's no Help URL. Previously, if there was no Help URL, the button would be enabled to show the tooltip, but nothing would happen when clicking it. Even with a Help URL, it could be confusing to have the Help button effectively serve two purposes.

Now the Help button doesn't show when there's no HelpURLAttribute. It's more space for the Label. And the Help button's tooltip now matches Unity's own Help button tooltips.